Moved Clock, MockClock and JavaClock to 'support' package
