+++ src/main/java/com/odeyalo/sonata/playlists/support/MockClock.java
@@ -1,4 +1,4 @@
-package com.odeyalo.sonata.playlists.service.tracks;
+package com.odeyalo.sonata.playlists.support;
 
 import org.jetbrains.annotations.NotNull;
